added like counter functionality
